requirements.txt
Jinja2~=3.1.3
aiogram~=3.7.0
python-dotenv~=1.0.1